changes 1

+++ unit/calculator.c
@@ -12,13 +12,15 @@ double subtract(double a, double b) {
 }
 
 double multiply(double a, double b) {
-    return a * b;
+    	return a * b;
 }
 
 double divide(double a, double b) {
     if (b == 0) return 0; 
     else if(a>b)
-    	return (a / b);
+    	return (int)(a / b);
     else
 	return (b/a);
+
 }
+
